requirements.txt
whitenoise==3.3.1
Django==1.11.13
xhtml2pdf==0.2.2
django_tabbed_admin==1.0.3
sorl_thumbnail==12.4.1
psycopg2
gunicorn
dj_database_url
django-stdimage